+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/FrameworkExtension.php
@@ -1272,26 +1272,31 @@ class FrameworkExtension extends Extension
             $fileRecorder('xml', dirname($reflClass->getFileName()).'/Resources/config/validation.xml');
         }
 
+        $this->registerComponentMapping($container, $fileRecorder, 'validation');
+
+        $this->registerMappingFilesFromConfig($container, $config, $fileRecorder);
+    }
+
+    private function registerComponentMapping(ContainerBuilder $container, $fileRecorder, $component)
+    {
         foreach ($container->getParameter('kernel.bundles_metadata') as $bundle) {
             $dirname = $bundle['path'];
 
             if (
-                $container->fileExists($file = $dirname.'/Resources/config/validation.yaml', false) ||
-                $container->fileExists($file = $dirname.'/Resources/config/validation.yml', false)
+                $container->fileExists($file = $dirname.'/Resources/config/'.$component.'.yaml', false) ||
+                $container->fileExists($file = $dirname.'/Resources/config/'.$component.'.yml', false)
             ) {
                 $fileRecorder('yml', $file);
             }
 
-            if ($container->fileExists($file = $dirname.'/Resources/config/validation.xml', false)) {
+            if ($container->fileExists($file = $dirname.'/Resources/config/'.$component.'.xml', false)) {
                 $fileRecorder('xml', $file);
             }
 
-            if ($container->fileExists($dir = $dirname.'/Resources/config/validation', '/^$/')) {
+            if ($container->fileExists($dir = $dirname.'/Resources/config/'.$component, '/^$/')) {
                 $this->registerMappingFilesFromDir($dir, $fileRecorder);
             }
         }
-
-        $this->registerMappingFilesFromConfig($container, $config, $fileRecorder);
     }
 
     private function registerMappingFilesFromDir($dir, callable $fileRecorder)
@@ -1387,10 +1392,36 @@ class FrameworkExtension extends Extension
 
         $loader->load('property_access.xml');
 
-        $container->getDefinition('property_accessor')->setPrivate(true);
+        $loaders = array();
+        $fileRecorder = function ($extension, $path) use (&$loaders) {
+            $definition = new Definition(in_array($extension, array('yaml', 'yml')) ? 'Symfony\Component\PropertyAccess\Mapping\Loader\YamlFileLoader' : 'Symfony\Component\PropertyAccess\Mapping\Loader\XmlFileLoader', array($path));
+            $definition->setPublic(false);
+            $loaders[] = $definition;
+        };
+
+        if (isset($config['enable_annotations']) && $config['enable_annotations']) {
+            if (!$this->annotationsConfigEnabled) {
+                throw new \LogicException('"enable_annotations" on property access cannot be set as Annotations support is disabled.');
+            }
+            $annotationLoader = new Definition(
+                'Symfony\Component\PropertyAccess\Mapping\Loader\AnnotationLoader',
+                array(new Reference('annotation_reader'))
+            );
+            $annotationLoader->setPublic(false);
+            $loaders[] = $annotationLoader;
+        }
+
+        $this->registerComponentMapping($container, $fileRecorder, 'property_accessor');
+
+        $this->registerMappingFilesFromConfig($container, $config, $fileRecorder);
+
+        $chainLoader = $container->getDefinition('property_access.mapping.chain_loader');
+
+        $chainLoader->replaceArgument(0, $loaders);
 
         $container
             ->getDefinition('property_accessor')
+            ->setPrivate(true)
             ->replaceArgument(0, $config['magic_call'])
             ->replaceArgument(1, $config['throw_exception_on_invalid_index'])
         ;