+++ src/Symfony/Component/PropertyAccess/PropertyAccessor.php
@@ -23,6 +23,8 @@ use Symfony\Component\PropertyAccess\Exception\InvalidArgumentException;
 use Symfony\Component\PropertyAccess\Exception\NoSuchIndexException;
 use Symfony\Component\PropertyAccess\Exception\NoSuchPropertyException;
 use Symfony\Component\PropertyAccess\Exception\UnexpectedTypeException;
+use Symfony\Component\PropertyAccess\Mapping\Factory\MetadataFactoryInterface;
+use Symfony\Component\PropertyAccess\Mapping\PropertyMetadata;
 
 /**
  * Default implementation of {@link PropertyAccessorInterface}.
@@ -30,6 +32,7 @@ use Symfony\Component\PropertyAccess\Exception\UnexpectedTypeException;
  * @author Bernhard Schussek <bschussek@gmail.com>
  * @author Kévin Dunglas <dunglas@gmail.com>
  * @author Nicolas Grekas <p@tchwork.com>
+ * @author Luis Ramón López <lrlopez@gmail.com>
  */
 class PropertyAccessor implements PropertyAccessorInterface
 {
@@ -119,6 +122,11 @@ class PropertyAccessor implements PropertyAccessorInterface
     const CACHE_PREFIX_PROPERTY_PATH = 'p';
 
     /**
+     * @internal
+     */
+    const CACHE_PREFIX_METADATA = 'm';
+
+    /**
      * @var bool
      */
     private $magicCall;
@@ -129,6 +137,14 @@ class PropertyAccessor implements PropertyAccessorInterface
      */
     private $cacheItemPool;
 
+    /**
+     * @var MetadataFactoryInterface
+     */
+    private $classMetadataFactory;
+
+    /**
+     * @var array
+     */
     private $readPropertyCache = array();
     private $writePropertyCache = array();
     private $propertyPathCache = array();
@@ -141,15 +157,17 @@ class PropertyAccessor implements PropertyAccessorInterface
      * Should not be used by application code. Use
      * {@link PropertyAccess::createPropertyAccessor()} instead.
      *
-     * @param bool                   $magicCall
-     * @param bool                   $throwExceptionOnInvalidIndex
-     * @param CacheItemPoolInterface $cacheItemPool
+     * @param bool                     $magicCall
+     * @param bool                     $throwExceptionOnInvalidIndex
+     * @param CacheItemPoolInterface   $cacheItemPool|null
+     * @param MetadataFactoryInterface $classMetadataFactory|null
      */
-    public function __construct($magicCall = false, $throwExceptionOnInvalidIndex = false, CacheItemPoolInterface $cacheItemPool = null)
+    public function __construct($magicCall = false, $throwExceptionOnInvalidIndex = false, CacheItemPoolInterface $cacheItemPool = null, MetadataFactoryInterface $classMetadataFactory = null)
     {
         $this->magicCall = $magicCall;
         $this->ignoreInvalidIndices = !$throwExceptionOnInvalidIndex;
         $this->cacheItemPool = $cacheItemPool instanceof NullAdapter ? null : $cacheItemPool; // Replace the NullAdapter by the null value
+        $this->classMetadataFactory = $classMetadataFactory;
     }
 
     /**
@@ -520,18 +538,20 @@ class PropertyAccessor implements PropertyAccessorInterface
         }
 
         if ($this->cacheItemPool) {
-            $item = $this->cacheItemPool->getItem(self::CACHE_PREFIX_READ.str_replace('\\', '.', $key));
+            $item = $this->cacheItemPool->getItem(self::CACHE_PREFIX_READ.$this->encodeCacheKey($key));
             if ($item->isHit()) {
                 return $this->readPropertyCache[$key] = $item->get();
             }
         }
+
+        $metadata = $this->getPropertyMetadata($class, $property);
 
         $access = array();
 
         $reflClass = new \ReflectionClass($class);
         $access[self::ACCESS_HAS_PROPERTY] = $reflClass->hasProperty($property);
         $camelProp = $this->camelize($property);
-        $getter = 'get'.$camelProp;
+        $getter = ($metadata && $metadata->getGetter()) ? $metadata->getGetter() : 'get'.$camelProp;
         $getsetter = lcfirst($camelProp); // jQuery style, e.g. read: last(), write: last($item)
         $isser = 'is'.$camelProp;
         $hasser = 'has'.$camelProp;
@@ -699,11 +719,13 @@ class PropertyAccessor implements PropertyAccessorInterface
         }
 
         if ($this->cacheItemPool) {
-            $item = $this->cacheItemPool->getItem(self::CACHE_PREFIX_WRITE.str_replace('\\', '.', $key));
+            $item = $this->cacheItemPool->getItem(self::CACHE_PREFIX_WRITE.$this->encodeCacheKey($key));
             if ($item->isHit()) {
                 return $this->writePropertyCache[$key] = $item->get();
             }
         }
+
+        $metadata = $this->getPropertyMetadata($class, $property);
 
         $access = array();
 
@@ -713,7 +735,7 @@ class PropertyAccessor implements PropertyAccessorInterface
         $singulars = (array) Inflector::singularize($camelized);
 
         if (is_array($value) || $value instanceof \Traversable) {
-            $methods = $this->findAdderAndRemover($reflClass, $singulars);
+            $methods = $this->findAdderAndRemover($reflClass, $singulars, $metadata);
 
             if (null !== $methods) {
                 $access[self::ACCESS_TYPE] = self::ACCESS_TYPE_ADDER_AND_REMOVER;
@@ -723,7 +745,8 @@ class PropertyAccessor implements PropertyAccessorInterface
         }
 
         if (!isset($access[self::ACCESS_TYPE])) {
-            $setter = 'set'.$camelized;
+            $setter = ($metadata && $metadata->getSetter()) ? $metadata->getSetter() : 'set'.$camelized;
+
             $getsetter = lcfirst($camelized); // jQuery style, e.g. read: last(), write: last($item)
 
             if ($this->isMethodAccessible($reflClass, $setter, 1)) {
@@ -742,7 +765,7 @@ class PropertyAccessor implements PropertyAccessorInterface
                 // we call the getter and hope the __call do the job
                 $access[self::ACCESS_TYPE] = self::ACCESS_TYPE_MAGIC;
                 $access[self::ACCESS_NAME] = $setter;
-            } elseif (null !== $methods = $this->findAdderAndRemover($reflClass, $singulars)) {
+            } elseif (null !== $methods = $this->findAdderAndRemover($reflClass, $singulars, $metadata)) {
                 $access[self::ACCESS_TYPE] = self::ACCESS_TYPE_NOT_FOUND;
                 $access[self::ACCESS_NAME] = sprintf(
                     'The property "%s" in class "%s" can be defined with the methods "%s()" but '.
@@ -819,11 +842,17 @@ class PropertyAccessor implements PropertyAccessorInterface
      *
      * @return array|null An array containing the adder and remover when found, null otherwise
      */
-    private function findAdderAndRemover(\ReflectionClass $reflClass, array $singulars)
+    private function findAdderAndRemover(\ReflectionClass $reflClass, array $singulars, PropertyMetadata $metadata = null)
     {
+        $fixedAdder = ($metadata && $metadata->getAdder()) ? $metadata->getAdder() : null;
+        $fixedRemover = ($metadata && $metadata->getRemover()) ? $metadata->getRemover() : null;
+        if ($fixedAdder && $fixedRemover) {
+            return array($fixedAdder, $fixedRemover);
+        }
+
         foreach ($singulars as $singular) {
-            $addMethod = 'add'.$singular;
-            $removeMethod = 'remove'.$singular;
+            $addMethod = $fixedAdder ?: 'add'.$singular;
+            $removeMethod = $fixedRemover ?: 'remove'.$singular;
 
             $addMethodFound = $this->isMethodAccessible($reflClass, $addMethod, 1);
             $removeMethodFound = $this->isMethodAccessible($reflClass, $removeMethod, 1);
@@ -832,6 +861,8 @@ class PropertyAccessor implements PropertyAccessorInterface
                 return array($addMethod, $removeMethod);
             }
         }
+
+        return null;
     }
 
     /**
@@ -922,5 +953,43 @@ class PropertyAccessor implements PropertyAccessorInterface
         }
 
         return $apcu;
+    }
+
+    /**
+     * Returns metadata associated with the property if it exists.
+     *
+     * @param $class
+     * @param $property
+     *
+     * @return null|PropertyMetadata
+     */
+    private function getPropertyMetadata($class, $property)
+    {
+        if ($this->cacheItemPool) {
+            $key = (false !== strpos($class, '@') ? rawurlencode($class) : $class).'..'.$property;
+            $item = $this->cacheItemPool->getItem(self::CACHE_PREFIX_METADATA.$this->encodeCacheKey($key));
+            if ($item->isHit()) {
+                return $item->get();
+            }
+        }
+
+        $metadata = null;
+        if ($this->classMetadataFactory && $classMetadata = $this->classMetadataFactory->getMetadataFor($class)) {
+            $metadata = $classMetadata->getMetadataForProperty($property);
+        }
+
+        return $metadata;
+    }
+
+    /**
+     * Escapes the key so it does not contains invalid characters.
+     *
+     * @param string $key
+     *
+     * @return string
+     */
+    private function encodeCacheKey($key)
+    {
+        return str_replace('\\', '.', $key);
     }
 }